restored rest of Surveys page markup
