+++ Decsys/ClientApp/src/app/admin/surveys/Surveys.js
@@ -1,19 +1,27 @@
 import React from "react";
-import { connect } from "react-redux";
-import { getSurveysSortOrder } from "./_selectors";
-import Survey from "./Survey";
+import { Typography, Grid, Row, Button, Alert } from "@smooth-ui/core-sc";
+import { InfoCircle, PlusCircle } from "styled-icons/fa-solid";
+import SurveyList from "./SurveyList";
 
-const SurveyList = ({ surveys, surveysSortOrder }) => (
-  <ul>
-    {surveysSortOrder.map(
-      id => !!surveys[id] && <Survey key={id} {...surveys[id]} />
-    )}
-  </ul>
+// TODO: Create New Survey wiring
+const Surveys = () => (
+  <Grid>
+    <Row my="3em" alignItems="center" justifyContent="space-between">
+      <Typography variant="h1">My Surveys</Typography>
+      <Button variant="secondary">
+        <PlusCircle size="1em" /> Create New Survey
+      </Button>
+    </Row>
+
+    <Row>
+      <Alert width="100%" variant="info">
+        <InfoCircle size="1em" /> Please don't forget to backup your surveys and
+        results to an external source.
+      </Alert>
+    </Row>
+
+    <SurveyList />
+  </Grid>
 );
-
-const Surveys = connect(state => ({
-  surveysSortOrder: getSurveysSortOrder(state),
-  surveys: state.data.surveys
-}))(SurveyList);
 
 export default Surveys;
